Split instruction_set implementations
